fix: eslint and move functions
